+++ src/settings/settings.service.ts
@@ -21,7 +21,7 @@ import copy from "fast-copy";
 import { CalendariumNotice } from "src/utils/notice";
 import { SyncBehavior } from "src/schemas";
 import {
-    isOlderVersion,
+    isOlder,
     MarkdownReason,
     shouldTransitionMarkdownSettings,
 } from "./settings.utils";
@@ -37,6 +37,7 @@ export default class SettingsService {
     #asking: boolean = false;
     #prompting: boolean = false;
     #saving: boolean = false;
+    deletedCalendars: Calendar[] = [];
 
     get app() {
         return this.plugin.app;
@@ -60,30 +61,6 @@ export default class SettingsService {
         this.app.workspace.onLayoutReady(() => (this.layoutReady = true));
         this.onLayoutReadyAndSettingsLoad(async () => {
             setTimeout(() => this.checkFCSettings(), 2000);
-            const permanentlyDelete: string[] = [];
-
-            console.debug(
-                `Calendarium: Checking deleted calendars for any to permanently delete.`
-            );
-            for (const calendar of this.#data.deletedCalendars) {
-                if (
-                    Date.now() - calendar.deletedTimestamp >
-                    7 * 24 * 60 * 60 * 1000
-                ) {
-                    permanentlyDelete.push(calendar.id);
-                }
-            }
-
-            if (permanentlyDelete.length) {
-                console.debug(
-                    `Calendarium: Found ${permanentlyDelete.length} deleted calendars more than 7 days old. Removing them.`
-                );
-                this.#data.deletedCalendars =
-                    this.#data.deletedCalendars.filter(
-                        (d) => !permanentlyDelete.includes(d.id)
-                    );
-                await this.saveData(this.#data);
-            }
         });
     }
     /**
@@ -121,65 +98,24 @@ export default class SettingsService {
     /**
      * This method is called whenever Obsidian detects that my data.json file has been modified.
      */
-    public onExternalSettingsChange = debounce(
-        async (): Promise<void> => {
-            // If I was the source of my data file change, just ignore this.
-            if (this.#saving) {
-                setTimeout(() => {
-                    this.#saving = false;
-                }, 500);
-                return;
-            }
-            if (this.syncPlugin._loaded) {
-                if (this.syncPlugin.instance.getStatus() !== "synced") {
-                    if (!this.#waitingOnSync) {
-                        this.#waitingOnSync = true;
-                        console.debug(
-                            "Calendarium: Obsidian Sync is actively syncing. Scheduling a reload after it completes."
-                        );
-                        const ref = this.syncPlugin.instance.on(
-                            "status-change",
-                            () => {
-                                if (
-                                    this.syncPlugin.instance.getStatus() !==
-                                    "synced"
-                                )
-                                    return;
-                                setTimeout(() => {
-                                    console.debug(
-                                        "Calendarium: Obsidian Sync finished. Performing reload."
-                                    );
-                                    this.#waitingOnSync = false;
-                                    this.onExternalSettingsChange();
-                                    this.syncPlugin.instance.offref(ref);
-                                }, 1000);
-                            }
-                        );
-                        this.plugin.registerEvent(ref);
-                    }
-                    return;
-                }
-            }
-            // If the user doesn't want their data synced, just ignore this.
-            if (this.#data.syncBehavior === "Never") {
-                console.debug(
-                    "Calendarium: Ignoring external data change event due to syncBehavior being 'Never'"
-                );
-                return;
-            }
-            // If the user wants it automatically synced, reload it.
-            if (this.#data.syncBehavior === "Always") {
-                console.debug(
-                    "Calendarium: Automatically reloading data due to syncBehavior being 'Always'"
-                );
-                await this.loadData(true);
-                return;
-            }
-            this.askToReload();
-        },
-        2000,
-        true
-    );
+    public async onExternalSettingsChange(): Promise<void> {
+        // If the user doesn't want their data synced, just ignore this.
+        if (this.#data.syncBehavior === "Never") {
+            console.debug(
+                "Calendarium: Ignoring external data change event due to syncBehavior being 'Never'"
+            );
+            return;
+        }
+        // If the user wants it automatically synced, reload it.
+        if (this.#data.syncBehavior === "Always") {
+            console.debug(
+                "Calendarium: Automatically reloading data due to syncBehavior being 'Always'"
+            );
+            await this.loadData(true);
+            return;
+        }
+        this.askToReload();
+    }
     private askToReload() {
         // If I am already asking, I shouldn't ask again.
         if (this.#asking) return;
@@ -340,8 +276,8 @@ export default class SettingsService {
         }
         return version;
     }
-    isOldVersion(older: Version, current: Version): boolean {
-        return isOlderVersion(older, current);
+    isOlder(older: Version, current: Version): boolean {
+        return isOlder(older, current);
     }
 
     /**
@@ -386,6 +322,8 @@ export default class SettingsService {
             );
             this.app.workspace.trigger("calendarium-updated");
         }
+
+        this.#saving = false;
     }
 
     /**
@@ -478,10 +416,7 @@ export default class SettingsService {
             this.#data.defaultCalendar = this.#data.calendars[0]?.id;
             this.plugin.watcher.start();
         }
-        this.#data.deletedCalendars.push({
-            ...copy(calendar),
-            deletedTimestamp: Date.now(),
-        });
+        this.deletedCalendars.push(calendar);
         await this.saveData(this.#data, true);
     }
 
@@ -646,18 +581,21 @@ export default class SettingsService {
             data.defaultCalendar = data.calendars[0].id;
             dirty = true;
         }
-        if (!data.deletedCalendars) {
-            data.deletedCalendars = [];
+        if ("deletedCalendars" in data) {
+            delete data.deletedCalendars;
             dirty = true;
         }
         /** Beta 29 */
         if (
-            this.isOldVersion(this.getDataVersion(data), {
-                major: 1,
-                minor: 0,
-                patch: 0,
-                beta: 29,
-            })
+            this.isOlder(
+                {
+                    major: 1,
+                    minor: 0,
+                    patch: 0,
+                    beta: 29,
+                },
+                this.getDataVersion(data)
+            )
         ) {
             data.paths = [];
             if (data.calendars.length) {
@@ -710,6 +648,11 @@ export default class SettingsService {
                     break;
                 }
             }
+
+            console.log(
+                "🚀 ~ file: settings.service.ts:664 ~ data.paths:",
+                data.paths
+            );
 
             if (data.calendars.length) {
                 data.inlineEventsTag =